+++ CHANGELOG.md
@@ -1,5 +1,14 @@
 # Changelog
 
+## [Unreleased]
+
+### New Features
+* Add workaround for Steam always setting `$LC_ALL` to `C` which prevented the correct game language from being selected
+
+### Bug Fixes
+* Properly parse proton when it's in `compatibilitytools.d` dir. Custom versions of proton, such as [Proton-GE], are stored here.
+* Fix `$XDG_CONFIG_DIR` not always being parsed properly
+
 ## [0.3.0] - 2020-12-28
 
 ### New Features
@@ -37,3 +46,4 @@ Initial release
 [0.2.0]: https://github.com/rfvgyhn/min-ed-launcher/compare/v0.1.1...v0.2.0
 [0.1.1]: https://github.com/rfvgyhn/min-ed-launcher/compare/v0.1.0...v0.1.1
 [0.1.0]: https://github.com/rfvgyhn/min-ed-launcher/compare/67d8c3f...v0.1.0
+[Proton-GE]: https://github.com/GloriousEggroll/proton-ge-custom